requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
datasets
shap
lime
opencv-python
deepface
mtcnn
torchvision
scikit-learn
matplotlib
seaborn
numpy
pandas
tqdm
Pillow
flask
flask-socketio
flask-cors
openai-whisper
pyttsx3
gtts
eventlet
mediapipe
